+++ lucene/test-framework/src/test/org/apache/lucene/search/TestBaseExplanationTestCase.java
@@ -89,7 +89,7 @@ public class TestBaseExplanationTestCase extends BaseExplanationTestCase {
       Explanation result = in.explain(context, doc);
       if (result.isMatch()) {
         if (q.breakExplainScores) {
-          result = Explanation.match(-1F * result.getValue(), "Broken Explanation Score", result);
+          result = Explanation.match(-1F * result.getValue().doubleValue(), "Broken Explanation Score", result);
         }
         if (q.toggleExplainMatch) {
           result = Explanation.noMatch("Broken Explanation Matching", result);
